# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WebEncode API
  description: |
    WebEncode is a distributed video encoding platform that provides VOD transcoding,
    live streaming management, and multi-platform publishing capabilities.
    
    ## Authentication
    All API endpoints require authentication via Bearer token in the Authorization header.
    Use the Auth Plugin to obtain tokens.
    
    ## Rate Limiting
    API requests are rate limited to 100 requests per minute per IP address.
    Exceeding this limit will result in a 429 Too Many Requests response.
  version: 1.0.0
  contact:
    name: WebEncode Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8090
    description: Local development
  - url: https://api.webencode.example.com
    description: Production
...
paths:
...
  /v1/restreams:
    get:
      tags: [Restreams]
      summary: List restream jobs
      responses:
        '200':
          description: List of restream jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestreamJob'
    post:
      tags: [Restreams]
      summary: Create restream job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestreamRequest'
      responses:
        '201':
          description: Restream job created
  /v1/restreams/{id}/start:
    post:
      tags: [Restreams]
      summary: Start a restream job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Restream started
  /v1/restreams/{id}/stop:
    post:
      tags: [Restreams]
      summary: Stop a restream job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Restream stopped
...
components:
...
  schemas:
...
    RestreamJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        input_url:
          type: string
        output_destinations:
          type: array
          items:
            type: object
        status:
          type: string
        created_at:
          type: string
          format: date-time
    CreateRestreamRequest:
      type: object
      properties:
        title:
          type: string
        input_url:
          type: string
        output_destinations:
          type: array
          items:
            type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []